Added tests in CitySearch
The first test ensures that no list (i.e., the suggestion list) is shown within the CitySearch component by default (notice that .not is stated before .toBeInTheDocument()).

The second test ensures that the suggestion list appears when the city input field “gains focus” (i.e., when the input field is clicked).